+++ src/main/java/com/gbtec/Main.java
@@ -1,9 +1,12 @@
 package com.gbtec;
 
 
+import com.gbtec.Controlsystem.GenerationController;
+
 public class Main {
 
     public static void main(String[] args) throws InterruptedException{
-        GenerationController generationController = new GenerationController(1,50,2);
+
+        GenerationController generationController = new GenerationController(0,100,2);
     }
-    }
+}
