+++ src/test/java/com/gbtec/CellTest.java
@@ -1,5 +1,6 @@
 package com.gbtec;
 
+import com.gbtec.Initialization.Cell;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.*;
