+++ src/test/java/com/gbtec/GenerationTest.java
@@ -1,6 +1,8 @@
 package com.gbtec;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import com.gbtec.Initialization.Generation;
 import org.junit.jupiter.api.Test;
 
 public class GenerationTest {
