+++ src/test/java/com/gbtec/NextGenerationTest.java
@@ -2,6 +2,9 @@ package com.gbtec;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
+import com.gbtec.Initialization.Generation;
+import com.gbtec.Initialization.NextGeneration;
+import com.gbtec.Visualization.GenerationPrinter;
 import org.junit.jupiter.api.Test;
